backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
yt-dlp>=2024.12.0
python-multipart==0.0.12
python-dotenv==1.0.1
slowapi==0.1.9
cachetools==5.5.0
aiofiles==24.1.0
httpx==0.27.2
